trying n as float

+++ final/sc/stream_compaction/compact_stream.cu
@@ -78,14 +78,14 @@ void compact_stream(void)
 // Use the GPU to compact the h_data stream 
 int compact_stream_on_device(float *result_d, float *h_data, unsigned int num_elements)
 {
-    int n = 0; // Number of elements in the compacted stream
-    int n_device = NULL;
+    float n = 0; // Number of elements in the compacted stream
+    float n_device = NULL;
     // device vectors
     float *result_device = NULL;
     float *h_device = NULL;
 
-    cudaMalloc((void**)&n, sizeof(int));
-    cudaMemcpy(n_device, n, sizeof(int), cudaMemcpyHostToDevice);
+    cudaMalloc((void**)&n, sizeof(float));
+    cudaMemcpy(n_device, n, sizeof(float), cudaMemcpyHostToDevice);
 
     cudaMalloc((void**)&result_device, num_elements*sizeof(float));
     cudaMemcpy(result_device, result_d, num_elements*sizeof(float), cudaMemcpyHostToDevice);
